Configure telemetry using Azure
If an EnvironmentCredential is configured, it is used to access AzureKeyVault where the connection string for the telemetry service is available. The same credential is used for publishing trace logs to Application Insights.

Also set up the MainWindow to appear sooner, so it looks like the app is launching. It sometimes still takes a long time for anything to show up.